+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.13.10
+
+* Trust plugins platform declaration when assigning tags.
+
+## 0.13.9+1
+
+* Fixed NPE in old Flutter plugin format detection.
+
 ## 0.13.9
 
 * Adds a tag for null safety.